requirements.txt
Flask>=3.0,<4.0
Flask-SQLAlchemy>=3.1,<4.0
Flask-Migrate>=4.0,<5.0
flask-limiter>=3.0,<4.0
flask-wtf>=1.2,<2.0
flask-cors>=4.0,<5.0
SQLAlchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
bcrypt>=4.0,<5.0
PyJWT>=2.8,<3.0
python-dotenv>=1.0,<2.0
itsdangerous>=2.0,<3.0
requests>=2.28,<3.0
gunicorn>=22.0,<23.0
google-generativeai>=0.3,<1.0
# Optional: Gmail API email sending
google-api-python-client>=2.0,<3.0
google-auth>=2.0,<3.0
# Optional: Azure Communication Services email
# azure-communication-email>=1.0,<2.0
# Optional: Stripe payment processing (leave uninstalled for mock billing mode)
# stripe>=7.0,<8.0
pytest>=8.0,<9.0
